to --to option R which outputs R
non-code lines are replaced with a single #
(this avoid creating trailing whitespace lintr errors on non-code lines)
also add json-block writer

This commit introduces a new writer, `json-block`, which serializes the
pandoc block at a specific line number to JSON.

A new `--line-number` command-line argument is added to specify the
target line. This argument is required when using `--to json-block`.